+++ src/main/java/seedu/duke/commands/ExitCommand.java
@@ -1,6 +1,6 @@
-package seedu.duke.Commands;
+package seedu.duke.commands;
 
-import seedu.duke.ProjectStuff.ProjectList;
+import seedu.duke.projectStuff.ProjectList;
 
 /**
  * Displays goodbye message when application is terminated by the user.
