+++ src/main/java/seedu/duke/commands/projectCommands/ProjectSelectCommand.java
@@ -1,21 +1,19 @@
-package seedu.duke.Commands.TaskCommands;
+package seedu.duke.commands.projectCommands;
 
-import seedu.duke.Commands.Command;
-import seedu.duke.ProjectStuff.ProjectList;
+import seedu.duke.commands.Command;
+import seedu.duke.projectStuff.ProjectList;
 
-public class TaskSelectCommand extends Command {
+public class ProjectSelectCommand extends Command {
 
     private int itemIndex;
-    private int projectIndex;
 
-    public TaskSelectCommand(int itemIndex, int projectIndex) {
+    public ProjectSelectCommand(int itemIndex) {
         this.itemIndex = itemIndex;
-        this.projectIndex = projectIndex;
     }
 
     public void executeCommand(ProjectList projects) {
         try {
-            projects.getProject(projectIndex).selectTask(itemIndex);
+            projects.selectProject(itemIndex);
         } catch (NumberFormatException | IndexOutOfBoundsException | NullPointerException e) {
             if (projects.getProjectList().size() == 0) {
                 System.out.println("I am empty!!!");
