+++ src/main/java/seedu/duke/commands/taskCommands/TaskDeleteCommand.java
@@ -1,7 +1,7 @@
-package seedu.duke.Commands.TaskCommands;
+package seedu.duke.commands.taskCommands;
 
-import seedu.duke.Commands.Command;
-import seedu.duke.ProjectStuff.ProjectList;
+import seedu.duke.commands.Command;
+import seedu.duke.projectStuff.ProjectList;
 
 public class TaskDeleteCommand extends Command {
