+++ src/main/java/seedu/duke/taskStuff/TaskList.java
@@ -1,4 +1,4 @@
-package seedu.duke.TaskStuff;
+package seedu.duke.taskStuff;
 
 import java.util.ArrayList;
